# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "业界优秀的通用管理框架基座（非业务系统）：RBAC 权限、菜单、部门、字典、操作日志、登录日志、数据范围、系统配置。\n清洁架构 handler → service → repository → model，复用泛型 Repository 基类。",
        "title": "GVA 通用管理框架基座 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/api",
    "paths": {
...
        "/system/login-log": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login-log"
                ],
                "summary": "登录日志分页列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户名/IP 关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "success/failed",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "起始时间",
                        "name": "startTime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "endTime",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login-log"
                ],
                "summary": "批量删除登录日志",
                "parameters": [
                    {
                        "description": "ID 列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.loginLogBatchDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    }
                }
            }
        },
        "/system/login-log/clear": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login-log"
                ],
                "summary": "清空登录日志",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    }
                }
            }
        },
        "/system/login-log/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login-log"
                ],
                "summary": "登录日志详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ProblemDetail"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login-log"
                ],
                "summary": "删除登录日志",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    }
                }
            }
        },
...
        "/system/operation-log": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operation-log"
                ],
                "summary": "操作日志分页列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户名/方法/路径关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态码",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "起始时间",
                        "name": "startTime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "endTime",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operation-log"
                ],
                "summary": "批量删除操作日志",
                "parameters": [
                    {
                        "description": "ID 列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.opLogBatchDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    }
                }
            }
        },
        "/system/operation-log/clear": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operation-log"
                ],
                "summary": "清空操作日志",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    }
                }
            }
        },
        "/system/operation-log/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operation-log"
                ],
                "summary": "操作日志详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ProblemDetail"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operation-log"
                ],
                "summary": "删除操作日志",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gva_internal_pkg_response.ApiResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
        "gva_internal_pkg_response.ProblemDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "应用层错误码（机器可读）",
                    "type": "string"
                },
                "detail": {
                    "description": "具体说明",
                    "type": "string"
                },
                "errors": {
                    "description": "字段级错误（表单校验）",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "description": "HTTP 状态码",
                    "type": "integer"
                },
                "title": {
                    "description": "简短摘要",
                    "type": "string"
                },
                "traceId": {
                    "type": "string"
                },
                "type": {
                    "description": "问题类型 URI",
                    "type": "string"
                }
            }
        },
...
        "internal_handler.loginLogBatchDeleteReq": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_handler.opLogBatchDeleteReq": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer {accessToken}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}